create new toast bar and restructure reusable notification styles
