+++ Projects/hCoreUI/Sources/Views/InfoCard.swift
@@ -3,13 +3,13 @@ import hCore
 
 public struct InfoCard: View {
     let text: String
-    let type: InfoCardType
+    let type: NotificationType
     @Environment(\.hInfoCardButtonConfig) var buttonsConfig
     @Environment(\.hInfoCardCustomView) var customContentView
 
     public init(
         text: String,
-        type: InfoCardType
+        type: NotificationType
     ) {
         self.text = text
         self.type = type
@@ -20,7 +20,7 @@ public struct InfoCard: View {
             VStack(spacing: .padding8) {
                 Image(uiImage: type.image)
                     .resizable()
-                    .foregroundColor(imageColor)
+                    .foregroundColor(type.imageColor)
                     .frame(width: 20, height: 20)
             }
             if let customContentView = customContentView {
@@ -30,7 +30,7 @@ public struct InfoCard: View {
             } else {
                 VStack(alignment: .leading) {
                     hText(text, style: .footnote)
-                        .foregroundColor(getTextColor)
+                        .foregroundColor(type.textColor)
                         .multilineTextAlignment(.leading)
                     if let buttonsConfig {
                         if buttonsConfig.count > 1 {
@@ -61,40 +61,9 @@ public struct InfoCard: View {
             }
         }
         .frame(maxWidth: .infinity, alignment: .leading)
-        .padding(.top, 14)
-        .padding(.bottom, .padding16)
-        .padding(.leading, .padding12)
-        .padding(.trailing, .padding16)
-        .modifier(InfoCardStyle(type: type))
-        .fixedSize(horizontal: false, vertical: true)
-    }
-
-    @hColorBuilder
-    var getTextColor: some hColor {
-        switch type {
-        case .info:
-            hSignalColor.Blue.text
-        case .attention:
-            hSignalColor.Amber.text
-        case .error:
-            hSignalColor.Red.text
-        case .campaign:
-            hSignalColor.Green.text
-        }
-    }
-
-    @hColorBuilder
-    var imageColor: some hColor {
-        switch type {
-        case .info:
-            hSignalColor.Blue.element
-        case .attention:
-            hSignalColor.Amber.element
-        case .error:
-            hSignalColor.Red.element
-        case .campaign:
-            hSignalColor.Green.element
-        }
+        .padding(.vertical, .padding12)
+        .padding(.horizontal, .padding16)
+        .modifier(NotificationStyle(type: type))
     }
 }
 
@@ -139,27 +108,6 @@ struct InfoCard_Previews: PreviewProvider {
 
                     }
             }
-        }
-    }
-}
-
-public enum InfoCardType {
-    case info
-    case attention
-    case error
-    case campaign
-
-    var image: UIImage {
-        switch self {
-        case .info:
-            return hCoreUIAssets.infoFilled.image
-        case .attention:
-            return hCoreUIAssets.warningTriangleFilled.image
-        case .error:
-            return hCoreUIAssets.warningTriangleFilled.image
-        case .campaign:
-            return hCoreUIAssets.campaignSmall.image
-
         }
     }
 }
@@ -228,47 +176,4 @@ extension View {
 public enum InfoCardLayoutStyle {
     case roundedRectangle
     case rectange
-}
-
-struct InfoCardStyle: ViewModifier {
-    let type: InfoCardType
-    @Environment(\.hInfoCardLayoutStyle) var layoutStyle
-    func body(content: Content) -> some View {
-        switch layoutStyle {
-        case .rectange:
-            content
-                .background(
-                    Rectangle()
-                        .fill(getBackgroundColor)
-                        .overlay(
-                            Rectangle()
-                                .strokeBorder(hBorderColor.primary, lineWidth: 0.5)
-                        )
-                )
-        case .roundedRectangle:
-            content
-                .background(
-                    Squircle.default()
-                        .fill(getBackgroundColor)
-                        .overlay(
-                            RoundedRectangle(cornerRadius: .cornerRadiusL)
-                                .strokeBorder(hBorderColor.primary, lineWidth: 0.5)
-                        )
-                )
-        }
-    }
-
-    @hColorBuilder
-    var getBackgroundColor: some hColor {
-        switch type {
-        case .info:
-            hSignalColor.Blue.fill
-        case .attention:
-            hSignalColor.Amber.fill
-        case .error:
-            hSignalColor.Red.fill
-        case .campaign:
-            hSignalColor.Green.fill
-        }
-    }
 }
